moved commands into separate folders
