+++ src/main/java/org/usfirst/frc4904/robot/commands/elevator/ElevatorControl.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands;
+package org.usfirst.frc4904.robot.commands.elevator;
 
 import java.util.function.BooleanSupplier;
