+++ src/main/java/org/usfirst/frc4904/robot/commands/manipulator/GroundIntake.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands;
+package org.usfirst.frc4904.robot.commands.manipulator;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
