+++ src/main/java/org/usfirst/frc4904/robot/commands/manipulator/ManipulatorRollerIn.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands;
+package org.usfirst.frc4904.robot.commands.manipulator;
 
 import org.usfirst.frc4904.robot.RobotMap;
 import org.usfirst.frc4904.robot.subsystems.Manipulator;
