+++ src/main/java/org/usfirst/frc4904/robot/commands/manipulator/ReadyGroundIntake.java
@@ -1,6 +1,8 @@
-package org.usfirst.frc4904.robot.commands;
+package org.usfirst.frc4904.robot.commands.manipulator;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
+import org.usfirst.frc4904.robot.commands.elevator.ElevatorUp;
+import org.usfirst.frc4904.robot.commands.elevator.ElevatorDown;
 
 public class ReadyGroundIntake extends CommandGroup {
 	public ReadyGroundIntake() {
